[PoC] Gmail Agent with MCP/requirements.txt
langgraph>=0.0.15
langchain>=0.1.0
google-generativeai>=0.3.0
python-telegram-bot>=20.7
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
aiohttp>=3.9.0